WebServer/requirements.txt
accelerate==0.31.0
aiohttp
aiosignal
alembic
annotated-types
anyio
asgiref
async-timeout
attrs
banal
beautifulsoup4
boto3
botocore
Brotli
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
cssselect
dataclasses-json
dataset
datasets
dill
Django
djangorestframework
dnspython
email_validator
exceptiongroup
fastapi
fastapi-cli
feedfinder2==0.0.4
feedparser
filelock
frozenlist
fsspec
google-api-core
google-auth
googleapis-common-protos
greenlet
h11
h2
hpack==4.0.0
httpcore
httpx
huggingface_hub
hyperframe
idna
importlib_metadata
importlib_resources
intel-openmp==2021.4.0
itsdangerous
jieba3k
Jinja2
jmespath
joblib
langchain
langsmith==0.0.92
lxml==4.9.4
Mako
markdown-it-py
MarkupSafe
marshmallow
marshmallow-enum
mdurl
mkl==2021.4.0
mpmath==1.3.0
multidict
multiprocess
mypy-extensions
networkx==3.2.1
newspaper3k
nltk
numexpr
numpy
openapi-schema-pydantic
orjson
packaging
pandas
pillow==10.2.0
proto-plus
protobuf==4.25.3
psutil==5.9.8
psycopg2
pyarrow==16.1.0
pyarrow-hotfix
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pyOpenSSL
PySocks
python-dateutil
python-dotenv
python-multipart
pytz
pyu2f
PyYAML
regex
requests
requests-file
rich
rsa
s3transfer
safetensors
sentencepiece==0.2.0
setuptools==69.5.1
sgmllib3k
shellingham
six
sniffio
soupsieve
SQLAlchemy
sqlparse
starlette
stringcase==1.2.0
sympy==1.12
tbb==2021.11.0
tenacity
tinysegmenter
tldextract
tokenizers
tqdm
typer==0.12.3
typer-slim==0.12.3
typing-inspect
typing_extensions
tzdata
ujson
urllib3
uvicorn
wheel==0.43.0
win-inet-pton
xxhash
yarl
zipp
